% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmesHousing.R
\docType{data}
\name{AmesHousing}
\alias{AmesHousing}
\title{Ames Housing data set}
\format{
\link{data.frame}
}
\source{
http://jse.amstat.org/v19n3/decock/AmesHousing.txt
}
\usage{
data(AmesHousing)
}
\description{
The benchmark data set Ames Housing in "raw" form.
}
\examples{
data(AmesHousing)
}
\references{
To be completed
}
\keyword{datasets}
