% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarsFramling.R
\docType{data}
\name{CarsFramling}
\alias{CarsFramling}
\title{Cars data set from Kary Främling's PhD thesis.}
\format{
\link{data.frame}
}
\usage{
data(CarsFramling)
}
\description{
Data collected from French car magazines in 1993.
}
\details{
The columns are:
\itemize{
\item Car model
\item Price (french francs)
\item Maximum power (hp)
\item Type of transmission (automatic/manual, scores 2/4)
\item Traction (front-/rear-/four-wheel driven, scores 1/2/3)
\item Car size (m3)
\item Trunk size (liters)
\item Weight (kg)
\item Top speed (km/h)
\item Acceleration 1000m with zero start speed (s)
\item Fuel consumption (l/100km)
\item Aesthetic score (1, 2, 3, 4 or 5), subjective
\item Fabricant’s image (1, 2, 3, 4 or 5), subjective
\item Score for standard equipment (continuous from 0 to 83).
\item Preference value (continuous from 0 to 100).
}
}
\examples{
data(CarsFramling)
}
\references{
Främling, Kary. Explaining Results of Neural Networks by Contextual Importance and Utility.
In: Robert Andrews and Joachim Diederich (eds.), Rules and networks:
Proceedings of the Rule Extraction from Trained Artificial Neural Networks Workshop,
AISB'96 conference, 1-2 April 1996. Brighton, UK, 1996.

Främling, Kary. Modélisation et apprentissage des préférences par
réseaux de neurones pour l'aide à la décision multicritère. PhD. thesis :
Institut National de Sciences Appliquées de Lyon, Ecole Nationale Supérieure
des Mines de Saint-Etienne, France, 1996. 209 p.
}
\keyword{datasets}
