% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.R
\name{ciu.explain}
\alias{ciu.explain}
\title{Explain CIU}
\usage{
ciu.explain(
  ciu,
  instance,
  ind.inputs.to.explain,
  in.min.max.limits = NULL,
  n.samples = 100,
  target.concept = NULL,
  target.ciu = NULL
)
}
\arguments{
\item{ciu}{blabla}

\item{instance}{blabla}

\item{ind.inputs.to.explain}{blabla}

\item{in.min.max.limits}{blabla}

\item{n.samples}{blabla}

\item{target.concept}{blabla}

\item{target.ciu}{blabla}
}
\value{
\code{ciu.result} object.
}
\description{
Explain CIU
}
\author{
Kary Främling
}
