% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeGraph.R
\name{ciu.get.sub.graph}
\alias{ciu.get.sub.graph}
\title{Extract subgraph from igraph}
\usage{
ciu.get.sub.graph(g, root_concept_names, relation = "feature-of", mode = "out")
}
\arguments{
\item{g}{\link[igraph:igraph]{igraph::igraph} object.}

\item{root_concept_names}{A root concept name, or an array of names.}

\item{relation}{The relation name to use for extraction.}

\item{mode}{The direction of the relation from the upper node to use for
extraction.}
}
\value{
An \link[igraph:igraph]{igraph::igraph}.
}
\description{
This function is mainly for extracting the concept tree for explanations,
starting from a root concept. If there are more than one root concepts, then
the corresponding subgraphs are extracted too.
}
