% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeGraph.R
\name{ciu.igraph.additive_attribution_to_ICs}
\alias{ciu.igraph.additive_attribution_to_ICs}
\title{Additive propagation of attributions to ICs}
\usage{
ciu.igraph.additive_attribution_to_ICs(
  graph,
  feature_attributions,
  attribute_name = "influence",
  normalize_to = NULL
)
}
\arguments{
\item{graph}{\link[igraph:igraph]{igraph::igraph} with the ICs to use. This should preferably
be a graph that only contains the actual explanation nodes, as extracted
e.g. by \link{ciu.get.sub.graph}.}

\item{feature_attributions}{Feature attribution values as an array that has
to include column names accessible by \code{names}.}

\item{attribute_name}{Node attribute name to use for the values.}

\item{normalize_to}{Array with minimal and maximal values to normalize the
values to. If set to NULL, then there won't be any normalisation.}
}
\value{
Updated \link[igraph:igraph]{igraph::igraph}.
}
\description{
This function is for making it possible to create Intermediate Concept (IC)
explanations/visualisations also with other methods than with CIU. This is
notably for Shapley values, where the influence value of an IC is the sum
of its constituents (which should add up to the difference between the
current output value and the baseline). In practice, it can of course be used
for any values but then there is no guarantee that there's a correspondence
between the additive result and the actual output value. Even for Shapley
values, there's no guarantee that it will be so due to approximated values
as produced by the \code{iml} package for instance. With SHAP, this should however
be guaranteed.
}
\examples{
\dontrun{
# This should not be run because it requires the use of one of the example
# files that are not a part of the package itself.
library(AmesHousing)
library(igraph)
library(ciu)
library(iml)
source("TestAmesKnowledgeGraph.R")
ames_pms <- create.CIU_Ames_for_Shiny()
kg <- ames_pms$partner_models$Default$graph
# Extract IC hierarchy from it
kg <- ciu.get.sub.graph(kg, "SalesPrice_def_voc")
# Get Shapley values with iml
inst_ind <- 16
ames_instance <- ames_pms$data[inst_ind,]
predict(ames_pms$model, ames_instance)
iml_model <- Predictor$new(ames_pms$model, data = ames_pms$data)
shapley_vals <- Shapley$new(iml_model, x.interest = ames_instance)
shapley_vals_vec <- setNames(shapley_vals$results$phi, shapley_vals$results$feature)
kg <- ciu.igraph.additive_attribution_to_ICs(kg, shapley_vals_vec)

# Create and show visNetwork for it
igraph::V(kg)$absinfluence <- abs(igraph::V(kg)$influence)
visnet <- ciu.plots.visNetwork(kg, node_area_attribute = "absinfluence",
  node_colors = ifelse(igraph::V(kg)$influence > 0, "steelblue", "firebrick"))
visnet

# With more appropriate tooltips:
tooltips <- paste0(V(kg)$name, "<br>&phi;: ", format(V(kg)$influence, digits = 3))
visnet <- ciu.plots.visNetwork(kg, node_area_attribute = "absinfluence",
  node_colors = ifelse(igraph::V(kg)$influence > 0, "steelblue", "firebrick"),
  node_tooltips = tooltips)
visnet

# Influence values for intermediate concepts can then be retrieved liek this:
fnames <- ciu.kg.get.child.features(kg, "Basement")
inf_values <- V(kg)[name \%in\% fnames]$influence
df <- data.frame(feature=fnames, phi=inf_values, sign=(inf_values>=0))
p <- ggplot(df) + geom_col(aes(x=reorder(feature, phi), y=phi, fill=sign)) +
  coord_flip() +
  labs(x ="", y = expression(phi)) + theme(legend.position = "none") +
  scale_fill_manual("legend", values = c("FALSE" = "firebrick", "TRUE" = "steelblue"))
print(p)
}
}
