% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeGraph.R
\name{ciu.igraph.explain}
\alias{ciu.igraph.explain}
\title{Set CIU values for graph nodes}
\usage{
ciu.igraph.explain(
  instance,
  CIU_object,
  graph,
  concept,
  out_index = 1,
  scale_to_IC = TRUE,
  IC_samples = 500,
  level = 1
)
}
\arguments{
\item{instance}{The explained instance (data.frame with one row).}

\item{CIU_object}{A CIU explainer object.}

\item{graph}{\link[igraph:igraph]{igraph::igraph} object with the Knowledge Graph.}

\item{concept}{Name of the root concept.}

\item{out_index}{Which output index of the model to use for CIU.}

\item{scale_to_IC}{Scale CI values to the CI of the parent Intermediate
Concept or use the "global" CI value, i.e. the one that is directly
relative to the output value.}

\item{IC_samples}{Number of samples to use for estimating CIU values}

\item{level}{The level in the hierarchy. This starts with one and is used
internally for the recursive traversal.}
}
\value{
An \link[igraph:igraph]{igraph::igraph} object
}
\description{
Sets CI and CU values for all "Intermediate Concept" nodes in the passed
graph object, starting from the given root concept. This is a
recursively used/usable function for going through the entire hierarchy down
to the basic features.
}
