% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeGraph.R
\name{ciu.kg.get.child.features}
\alias{ciu.kg.get.child.features}
\title{Get child nodes based on specific relation}
\usage{
ciu.kg.get.child.features(g, node_name, relation = "feature-of")
}
\arguments{
\item{g}{\link[igraph:igraph]{igraph::igraph} graph.}

\item{node_name}{Name of root node.}

\item{relation}{Relation name to use.}
}
\value{
An array of child node names.
}
\description{
Get all child nodes of the node with name "node_name" and whose edges
have a "relation" attribute with the given name.
}
