% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{ciu.launch.social_xai.shiny}
\alias{ciu.launch.social_xai.shiny}
\title{Launch the Social XAI Shiny app}
\usage{
ciu.launch.social_xai.shiny(parameters, ...)
}
\arguments{
\item{parameters}{\link{list} that has to have at least the following elements:
\itemize{
\item model: The model to explain.
\item data: \link{data.frame} with the actual data.
\item target_variable: Name of target variable column.
\item ciu: CIU explainer.
\item partner_models: \link{list} of "partner models", where each model has at least
\item NO: graph: Default knowledge graph that defines vocabulary etc.
}}

\item{...}{Passed directly to shinyApp().}
}
\value{
Void
}
\description{
Launch the Social XAI Shiny app
}
