% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartnerModel.R
\name{ciu.partnermodel.new}
\alias{ciu.partnermodel.new}
\title{Create a new partner model}
\usage{
ciu.partnermodel.new(name, root_concepts, graph)
}
\arguments{
\item{name}{Name of the model.}

\item{root_concepts}{Names of root concepts/nodes in the graph.}

\item{graph}{Knowledge Graph \link[igraph:igraph]{igraph::igraph} of concepts and relations
to use for explanations.}
}
\value{
A \link{list} that is also of class "PartnerModel".
}
\description{
Create a new partner model
}
