% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.R
\name{ciu.pie}
\alias{ciu.pie}
\title{ciu.pie}
\usage{
ciu.pie(
  ciu,
  instance,
  ind.inputs = NULL,
  ind.output = 1,
  in.min.max.limits = NULL,
  n.samples = 100,
  neutral.CU = 0.5,
  show.input.values = TRUE,
  concepts.to.explain = NULL,
  target.concept = NULL,
  target.ciu = NULL,
  color.ramp.below.neutral = NULL,
  color.ramp.above.neutral = NULL,
  sort = NULL,
  decreasing = FALSE,
  main = NULL,
  ...
)
}
\arguments{
\item{ciu}{blabla}

\item{instance}{blabla}

\item{ind.inputs}{blabla}

\item{ind.output}{blabla}

\item{in.min.max.limits}{blabla}

\item{n.samples}{blabla}

\item{neutral.CU}{blabla}

\item{show.input.values}{blabla}

\item{concepts.to.explain}{blabla}

\item{target.concept}{blabla}

\item{target.ciu}{blabla}

\item{color.ramp.below.neutral}{blabla}

\item{color.ramp.above.neutral}{blabla}

\item{sort}{blabla}

\item{decreasing}{blabla}

\item{main}{blabla}

\item{...}{blabla}
}
\value{
blabla
}
\description{
ciu.pie
}
\author{
Kary Främling
}
