% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeGraph.R
\name{ciu.plot.knowledgegraph}
\alias{ciu.plot.knowledgegraph}
\title{Plot concept tree using basic \link[igraph:igraph]{igraph::igraph} plotting}
\usage{
ciu.plot.knowledgegraph(
  g,
  root_concept,
  main = "Concept hierarchy",
  circular = FALSE,
  ...
)
}
\arguments{
\item{g}{The \link[igraph:igraph]{igraph::igraph}.}

\item{root_concept}{Name of the node to use as the top level.}

\item{main}{Plot title.}

\item{circular}{Plot with circular layout or not?}

\item{...}{Remaining parameters are passed to \code{plot} function.}
}
\value{
Void
}
\description{
This is mainly a convenience function for plotting trees "sideways", which
tends to be more readable. Also, nodes of type "Feature" are plotted in red,
all others in blue. The \link[igraph:igraph]{igraph::igraph} should be a tree structure for this
to really make sense.
}
\examples{
library(ciu)
library(igraph)

# Create a graph with 10 nodes using random graph structure (Erdos-Renyi model)
g <- make_tree(7, children = 2, mode = "out")
V(g)[1:3]$type <- "Intermediate concept"
V(g)[4:7]$type <- "Feature"
ciu.plot.knowledgegraph(g, "1")
}
