% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.plots.R
\name{ciu.plots.visNetwork}
\alias{ciu.plots.visNetwork}
\title{Make a visNetwork for the graph}
\usage{
ciu.plots.visNetwork(
  graph,
  node_area_attribute = NULL,
  node_color_attribute = NULL,
  node_colors = NULL,
  node_tooltips = NULL,
  hierarchical_layout = FALSE,
  default_radius = 20,
  min_radius = 0,
  max_radius = 50,
  default_node_color = "lightblue",
  default_node_border_color = "darkblue",
  node_color_palette = colorRampPalette(c("red", "yellow", "darkgreen"))(100)
)
}
\arguments{
\item{graph}{\link[igraph:igraph]{igraph::igraph} object.}

\item{node_area_attribute}{Name of node attribute to use for node area.}

\item{node_color_attribute}{Name of node attribute to use for node color.}

\item{node_colors}{Array of colors to use directly as node colors.}

\item{node_tooltips}{Array of tooltips to use for nodes instead of the
default ones.}

\item{hierarchical_layout}{Use hierarchical layout instead of default one?}

\item{default_radius}{Default radius for nodes.}

\item{min_radius}{Minimal radius allowed for nodes.}

\item{max_radius}{Maximal radius allowed for nodes.}

\item{default_node_color}{Default color for nodes.}

\item{default_node_border_color}{Default border color for nodes.}

\item{node_color_palette}{Color palette to use for node color.}
}
\value{
A \link[visNetwork:visNetwork]{visNetwork::visNetwork}.
}
\description{
Creates a visNetwork from an igraph and sets the visual attributes according
to the specified attribute.
}
\examples{
library(igraph)
library(visNetwork)
library(ciu)

# Create a graph with 10 nodes
# Using a random graph structure (Erdos-Renyi model)
g <- sample_gnp(10, 0.3)
# Add node attributes (CI, CU, influence)
V(g)$CI <- runif(10, min = 0, max = 1)
V(g)$CU <- runif(10, min = 0, max = 1)
V(g)$influence <- runif(10, min = -1, max = 1)
ciu.plots.plot_igraph(g) # No sizing, colors
ciu.plots.visNetwork(g, node_area_attribute = "CI", node_color_attribute = "CU")

# Influence plot
V(g)$absinfluence <- abs(V(g)$influence)
ciu.plots.visNetwork(g, node_area_attribute = "absinfluence",
  node_colors = ifelse(V(g)$influence > 0, "steelblue", "firebrick"))

# Add navigation buttons, just as an example
vn <- ciu.plots.visNetwork(g, node_area_attribute = "CI", node_color_attribute = "CU")
vn \%>\% visInteraction(navigationButtons = TRUE)

# Adding a legend color bar is more demanding and maybe not such a great
# idea. But it can be done through raw HTML:
library(htmltools)
ciu.plots.visNetwork_with_legend(vn)
# Legend and navigation buttons:
ciu.plots.visNetwork_with_legend(vn \%>\% visInteraction(navigationButtons = TRUE))
}
