% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.plots.R
\name{ciu.plots.visNetwork_CIU}
\alias{ciu.plots.visNetwork_CIU}
\title{Make a visNetwork for the igraph using CIU attributes}
\usage{
ciu.plots.visNetwork_CIU(graph, use.influence = FALSE, ...)
}
\arguments{
\item{graph}{CIU-initialized \link[igraph:igraph]{igraph::igraph}.}

\item{use.influence}{Use CI&CU values or influence?}

\item{...}{Parameters to pass to \link{ciu.plots.visNetwork}.}
}
\value{
A \link[visNetwork:visNetwork]{visNetwork::visNetwork}.
}
\description{
Mainly a CIU shortcut function for \link{ciu.plots.visNetwork}, which
allows for switching between CI&CU versus Contextual influence visualisation.
}
\examples{
library(ciu)
library(igraph)
library(visNetwork)

# Create a graph with 10 nodes using random graph structure (Erdos-Renyi model)
g <- sample_gnp(10, 0.3)
# Add node attributes (CI, CU, influence)
V(g)$CI <- runif(10, min = 0, max = 1)
V(g)$CU <- runif(10, min = 0, max = 1)
V(g)$influence <- runif(10, min = -1, max = 1)
ciu.plots.visNetwork_CIU(g)

}
