% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.plots.R
\name{ciu.plots.visNetwork_with_legend}
\alias{ciu.plots.visNetwork_with_legend}
\title{Create HTML plot with visNetwork and legend}
\usage{
ciu.plots.visNetwork_with_legend(vn, legend_title = "")
}
\arguments{
\item{vn}{\link[visNetwork:visNetwork]{visNetwork::visNetwork} object.}

\item{legend_title}{Title for the legend.}
}
\value{
A "browsable" object.
}
\description{
For the moment, this function just adds a legend with the default CU colors.
}
