% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextualImportanceUtility.R
\name{ciu.relative}
\alias{ciu.relative}
\title{Calculate relative CIU of a sub-concept/input relative to an intermediate
concept (or output).}
\usage{
ciu.relative(sub.ciu.result, sup.ciu.result)
}
\arguments{
\item{sub.ciu.result}{ciu.result object of sub-concept/input.}

\item{sup.ciu.result}{ciu.result object of intermediate concept/output.}
}
\description{
Calculate relative CIU of a sub-concept/input relative to an intermediate
concept (or output). The parameters must be of class "ciu.result" or a
data.frame with compatible columns.
}
