% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.R
\name{ciu.to.CIU}
\alias{ciu.to.CIU}
\title{Create \code{CIU} object from \code{ciu} object.}
\usage{
ciu.to.CIU(ciu)
}
\arguments{
\item{ciu}{blabla}
}
\value{
CIU object
}
\description{
Create \code{CIU} object from \code{ciu} object.
}
