% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeGraph.R
\name{ciu.voc.from.graph}
\alias{ciu.voc.from.graph}
\title{Get CIU vocabulary from Knowledge Graph}
\usage{
ciu.voc.from.graph(g, column_names)
}
\arguments{
\item{g}{\link[igraph:igraph]{igraph::igraph} object. The names of the leaves have to match
with the column names in \code{column_names}}

\item{column_names}{Array of column names in the data.}
}
\value{
A CIU vocabulary (a \link{list})
}
\description{
Make a CIU-compatible vocabulary based on the graph and data column names
Remember to remove the target value column(s) from the passed data!
}
