% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.R
\name{ciu.barplot}
\alias{ciu.barplot}
\title{ciu.barplot.ciu}
\usage{
ciu.barplot(
  ciu,
  instance,
  ind.inputs = NULL,
  ind.output = 1,
  in.min.max.limits = NULL,
  n.samples = 100,
  neutral.CU = 0.5,
  show.input.values = TRUE,
  concepts.to.explain = NULL,
  target.concept = NULL,
  target.ciu = NULL,
  color.ramp.below.neutral = NULL,
  color.ramp.above.neutral = NULL,
  use.influence = FALSE,
  influence.minmax = c(-1, 1),
  sort = NULL,
  decreasing = FALSE,
  main = NULL,
  xlab = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{ciu}{\code{ciu} object.}

\item{instance}{Instance to explain. See \code{\link{ciu.explain}}.}

\item{ind.inputs}{\code{\link{vector}} of indices for the inputs to be
included in the plot. If NULL then all inputs will be included.}

\item{ind.output}{Index of output to be explained.}

\item{in.min.max.limits}{See \code{\link{ciu.explain}}.}

\item{n.samples}{See \code{\link{ciu.explain}}.}

\item{neutral.CU}{Indicates when the Contextual Utility is considered
to be "negative". The default value of 0.5 seems quite logical for most cases.}

\item{show.input.values}{Include input values after input labels or not.
Default is TRUE.}

\item{concepts.to.explain}{List of concepts to use in the plot, as defined
by vocabulary provided as argument to \link{ciu.new}. If \code{ind.inputs=NULL},
then use \code{concepts.to.explain} instead. If both are \code{NULL}, then use all inputs.}

\item{target.concept}{See \link{ciu.explain}.}

\item{target.ciu}{See \link{ciu.explain}.}

\item{color.ramp.below.neutral}{Color ramp function as returned by function
\code{colorRamp()}. Default color ramp is from red3 to yellow.}

\item{color.ramp.above.neutral}{Color ramp function as returned by function
\code{colorRamp()}. Default colorramp is from yellow to darkgreen.}

\item{use.influence}{Plot using "influence" rather than CIU, i.e. a
LIME-like barplot. Default is FALSE.}

\item{sort}{NULL, "CI" or "CU". No sorting by default, other options are
sorting by CI or CU.}

\item{decreasing}{Set to TRUE for decreasing sort.}

\item{main}{Usual plot parameter, possible to override default one if needed.}

\item{xlab}{Usual plot parameter, possible to override default one if needed.}

\item{xlim}{Usual plot parameter, possible to override default one if needed.}

\item{...}{Other graphical parameters to pass to \link[base:plot]{base::plot}}
}
\value{
"void", i.e. whatever happens to be result of last instruction.
}
\description{
ciu.barplot.ciu
}
\seealso{
\link{ggplot.col.ciu}

\link{pie.ciu}

\link{ciu.new}

\link{ciu.explain}
}
\author{
Kary Främling
}
