% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextualInfluence.R
\name{ciu.contextual.influence}
\alias{ciu.contextual.influence}
\title{Get Contextual influence values}
\usage{
ciu.contextual.influence(
  ciu.result = NULL,
  CI = NULL,
  CU = NULL,
  neutral.CU = 0.5
)
}
\arguments{
\item{ciu.result}{\link{ciu.result} object. If NULL, then CI and CU values must be
provided.}

\item{CI}{CI value(s).}

\item{CU}{CU value(s).}

\item{neutral.CU}{Neutral CU value(s). Default is 0.5}
}
\value{
Contextual influence value(s). Value or vector.
}
\description{
Contextual influence is calculated from CI and CU values, using a baseline
that is considered a "neutral" CU value.
}
\author{
Kary Främling
}
