% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.ggplot.col.R
\name{ciu.ggplot.col}
\alias{ciu.ggplot.col}
\title{ciu.ggplot.col}
\usage{
ciu.ggplot.col(
  ciu,
  instance,
  ind.inputs = NULL,
  output.names = NULL,
  in.min.max.limits = NULL,
  n.samples = 100,
  neutral.CU = 0.5,
  show.input.values = TRUE,
  concepts.to.explain = NULL,
  target.concept = NULL,
  target.ciu = NULL,
  low.color = "red",
  mid.color = "yellow",
  high.color = "darkgreen",
  use.influence = FALSE,
  influence.minmax = c(-1, 1),
  sort = NULL,
  decreasing = FALSE,
  main = NULL
)
}
\arguments{
\item{ciu}{blabla}

\item{instance}{blabla}

\item{ind.inputs}{blabla}

\item{output.names}{blabla}

\item{in.min.max.limits}{blabla}

\item{n.samples}{blabla}

\item{neutral.CU}{blabla}

\item{show.input.values}{blabla}

\item{concepts.to.explain}{blabla}

\item{target.concept}{blabla}

\item{target.ciu}{blabla}

\item{low.color}{blabla}

\item{mid.color}{blabla}

\item{high.color}{blabla}

\item{use.influence}{Plot "influence" instead of CI and CU.}

\item{influence.minmax}{Range to use for influence.}

\item{sort}{blabla}

\item{decreasing}{blabla}

\item{main}{blabla}
}
\value{
blabla
}
\description{
ciu.ggplot.col
}
\author{
Kary Främling
}
