% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaExplain.R
\name{ciu.list.to.frame}
\alias{ciu.list.to.frame}
\title{ciu.list.to.frame}
\usage{
ciu.list.to.frame(ciu.list, out.ind = 1)
}
\arguments{
\item{ciu.list}{\link{list} of ciu.result objects.}

\item{out.ind}{Index of output to extract.}
}
\value{
\link{data.frame} with same columns as ciu.result object but with one row
per input feature.
}
\description{
Convert \link{list} of ciu.result objects into corresponding \link{data.frame} for
given output.
}
\author{
Kary Främling
}
