% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.R
\name{ciu.plot.3D}
\alias{ciu.plot.3D}
\title{ciu.plot.3D}
\usage{
ciu.plot.3D(
  ciu,
  instance,
  ind.inputs,
  ind.output,
  in.min.max.limits = NULL,
  n.points = 40,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  zlim = NULL,
  ...
)
}
\arguments{
\item{ciu}{\code{ciu} object.}

\item{instance}{blabla}

\item{ind.inputs}{blabla}

\item{ind.output}{blabla}

\item{in.min.max.limits}{blabla}

\item{n.points}{blabla}

\item{main}{blabla}

\item{xlab}{blabla}

\item{ylab}{blabla}

\item{zlab}{blabla}

\item{zlim}{blabla}

\item{...}{blabla}
}
\value{
blabla
}
\description{
ciu.plot.3D
}
\author{
Kary Främling
}
