% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciu.R
\name{ciu.plot}
\alias{ciu.plot}
\title{ciu.plot}
\usage{
ciu.plot(
  ciu,
  instance,
  ind.input,
  ind.output,
  in.min.max.limits = NULL,
  n.points = 40,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{ciu}{\code{ciu} object.}

\item{instance}{blabla}

\item{ind.input}{blabla}

\item{ind.output}{blabla}

\item{in.min.max.limits}{blabla}

\item{n.points}{blabla}

\item{main}{blabla}

\item{xlab}{blabla}

\item{ylab}{blabla}

\item{ylim}{blabla}

\item{...}{blabla}
}
\value{
blabla
}
\description{
ciu.plot
}
\author{
Kary Främling
}
